edit SnackMachine for correct EVA implementation

+++ noseryoung/ch/src/Snack.java
@@ -3,12 +3,12 @@ import java.util.concurrent.atomic.AtomicInteger;
 public class Snack {
     private String name;
     private int count;
-    private int price;
+    private float price;
     private int initialAmount;
     private final int id;
     private static AtomicInteger idCounter = new AtomicInteger(0);
 
-    public Snack(String name, int initialAmount, int price){
+    public Snack(String name, int initialAmount, float price){
        this.name= name;
        this.initialAmount = initialAmount;
        this.count = initialAmount;
@@ -36,11 +36,11 @@ public class Snack {
         this.name = name;
     }
 
-    public int getPrice() {
+    public float getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(float price) {
         this.price = price;
     }
